chore: move renameWidget function to widgetUtil.ts
